requirements.txt
openai>=1.30.0
requests>=2.31.0
notion-client>=2.2.1
python-dotenv>=1.0.0
